Patch symbol loading issue on linux. Tentatively turn on tests for linux (tests run locally, but have had problems running in the build).

+++ packages/light-source/addon/light-source/ls/SceneNode.cc
@@ -192,9 +192,9 @@ void SceneNode::InsertBefore(const CallbackInfo& info) {
     }
 
     auto before{ SceneNode::QueryInterface(info[1]) };
-    auto beforeIndex{ before ? this->GetChildIndex(before) : YGNodeGetChildCount(this->ygNode) };
+    auto beforeIndex{ before ? this->GetChildIndex(before) : static_cast<int32_t>(YGNodeGetChildCount(this->ygNode)) };
 
-    if (beforeIndex < 0u) {
+    if (beforeIndex < 0) {
         throw Error::New(env, "insertBefore: before argument is not a child");
     }
